fix: #665 validateTime mark invalid immediately if value is null
