--- src/app/app.component.spec.ts
@@ -1,20 +1,0 @@
-import { async, TestBed } from '@angular/core/testing';
-
-import { AppComponent } from './app.component';
-import { AppModule } from './app.module';
-
-describe('AppComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [AppModule],
-    }).compileComponents();
-  }));
-
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    fixture.detectChanges();
-    expect(app).toBeObject();
-    expect(app instanceof AppComponent).toBeTrue();
-  });
-});
