+++ src/app/bugs/bugs.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { FormBuilder, FormGroup } from '@angular/forms';
+
+@Component({
+  templateUrl: './bugs.component.html',
+})
+export class BugsComponent {
+  public bugsForm: FormGroup;
+  public constructor(private formBuilder: FormBuilder) {
+    this.bugsForm = this.formBuilder.group({ MonStart: [] });
+  }
+}
