+++ src/app/error/error.component.html
@@ -1,0 +1,16 @@
+<div class="container">
+  <div class="row">
+    <div class="col-12">
+      <div class="mat-grid-wr">
+        <mat-grid-list cols="1" rowHeight="55px">
+          <mat-grid-tile>
+            <mat-toolbar>
+              <span>Error!</span>
+            </mat-toolbar>
+          </mat-grid-tile>
+          <a routerLink="/">Back to showcase</a>
+        </mat-grid-list>
+      </div>
+    </div>
+  </div>
+</div>
