+++ src/app/error/error.component.ts
@@ -1,0 +1,6 @@
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: './error.component.html',
+})
+export class ErrorComponent {}
