+++ src/index.html
@@ -5,6 +5,7 @@
   <meta name="keywords" content="Development, JavaScript, TypeScript, Angular, CSS, SASS, HTML, WEB, GitHub, ng2Mask"/>
   <meta name="author" content=""/>
   <title>ng2Mask-Plugin - A Angular2 plugin to make field masks.</title>
+  <base href="/" />
   <link rel="stylesheet" href="http://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.11.0/styles/default.min.css">
 </head>
 <body>
