+++ src/karma.conf.js
@@ -32,10 +32,10 @@ module.exports = function(config) {
         },
         // thresholds per file
         each: {
-          statements: 5,
-          lines: 5,
-          branches: 5,
-          functions: 5,
+          statements: 0,
+          lines: 0,
+          branches: 0,
+          functions: 0,
         },
       },
     },
